requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]
python-jose
pydantic[email]
bcrypt==4.0.1
PyJWT
